servergles: don't put pointers in place of received offsets
OK we had the room in 32bit mode, but that required lots of casting, and
we already have too much of them already.  And besides, that just failed
to work on 64bit platforms.

Make it explicit that the offsets in the protocol are really that and not
string pointers as in the GL API.