serveregl: report EGL_VERSION/VENDOR advertising us and implementation level
The spec dictates that we must tell the same version as eglInitialize
does here.  We're not yet conforming to 1.0, and would like not to
tell more lies than necessary.  But glmark2 unconditionally uses
eglBindAPI from 1.2 and eglGetCurrentContext from 1.4, and getting 1.0
for EGL_VERSION causes its GL wrapper to get NULL for those function,
and to segfault after not checking them...