recvr: (temporarily) enlarge fifo packet size to 512MB
This is enough for many demos (including `glmark2 -b jellyfish`), but will
still not allow more complex ones (eg. `glmark2 -b terrain` and its 4.5MB
buffer).

Reduce the number of packets in FIFO from 1024 to 64 so the overall size
only grows by a factor of 16 not 256.  That's still a large number of
packets to live in a queue, likely only necessary when server uses software
rendering.