GLES: protect glBufferSubData against NULL pointer
NULL is not apparently valid for this call, but we have to protect ourselves
from following one.  Probably we should error out but the spec does not
mandate that, so let's pass it on to the server and hope.  Anyway if it
does not, we'll soon have process isolation so the impact will become minimal.