EGL: make eglGetProcAddress call on server too, and eventually return NULL
That ought not to be necessary, as version checks and extension declarations
ought to protect us from well-behaved apps, but let's be safe.

But foremost, this call will trigger caching of the function pointers
for extensions on server side so calling them will be smooth.

As a bonus, we now get warned when client app calls eglGetProcAddress
for a function supported on the server GL stack but not in GLS.