EGL: add (preliminary) support for multiple windows
Instead of enforcing use of a single window, we now track a remote2local
mapping of X11 windows, so relevant operations can apply to the proper
window.  With the current window-(non-)management implementation, however, we
don't track window destruction - but the target is using Qubes windows.

Steals associative list implementation from qubes-guid.  Not ideal, we'll
have to find better.