transport: isolate TCP code: tport_read() in recvr_handle_packet()
Doing that, we stop using MSG_PEEK, for which no similar
functionnality is available at least with pipes.  It was comfortable
to simplify recvr code, but to be generic we must use read() and a bit
of gymnastics instead.

Note that despite what the read(2) manpage says, reading a size of zero
seems to block in some cases, so change do/while() to plain while(), which
is also more efficient.

FIXME: server produces a stream of "transport socket closed" only stopping
when we get to "FIFO full"